Mixin fixes

+++ src/main/java/virtuoel/pehkui/mixin/compat116plus/compat118minus/LookTargetUtilMixin.java
@@ -22,7 +22,7 @@ public class LookTargetUtilMixin
 {
 	@Dynamic
 	@ModifyExpressionValue(method = MixinConstants.GIVE_TO_VEC3D, at = @At(value = "CONSTANT", args = "floatValue=0.3F", ordinal = 0))
-	private static double pehkui$give$offset(double value, LivingEntity entity, ItemStack stack, Vec3d targetLocation)
+	private static float pehkui$give$offset(float value, LivingEntity entity, ItemStack stack, Vec3d targetLocation)
 	{
 		final float scale = ScaleUtils.getEyeHeightScale(entity);
